Organize examples into step packages
