+++ app/src/main/kotlin/org/example/step1/Step1_Variables.kt
@@ -1,4 +1,4 @@
-package org.example
+package org.example.step1
 
 fun main() {
     val name = "Kotlin" // 변경 불가
