+++ app/src/main/kotlin/org/example/step10/Step10_DataClassEquals.kt
@@ -1,4 +1,4 @@
-package org.example
+package org.example.step10
 
 data class User(val id: Int, val name: String)
