+++ app/src/main/kotlin/org/example/step11/Step11_Sealed.kt
@@ -1,4 +1,4 @@
-package org.example
+package org.example.step11
 
 // 봉인 클래스
 sealed class Result
