+++ app/src/main/kotlin/org/example/step12/Step12_Enum.kt
@@ -1,4 +1,4 @@
-package org.example
+package org.example.step12
 
 enum class Direction {
     NORTH, SOUTH, EAST, WEST
