+++ app/src/main/kotlin/org/example/step14/Step14_ScopeFunctions.kt
@@ -1,4 +1,4 @@
-package org.example
+package org.example.step14
 
 data class Person(var name: String, var age: Int)
