+++ app/src/main/kotlin/org/example/step4/Step4_Control.kt
@@ -1,4 +1,4 @@
-package org.example
+package org.example.step4
 
 fun main() {
     val score = 85
