+++ app/src/main/kotlin/org/example/step6/Step6_NullSafety.kt
@@ -1,4 +1,4 @@
-package org.example
+package org.example.step6
 
 fun main() {
     // Nullable 선언
