+++ app/src/main/kotlin/org/example/step7/Step7_Lambda.kt
@@ -1,4 +1,4 @@
-package org.example
+package org.example.step7
 
 fun main() {
     // 람다식 정의
